requirements.txt
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-community
fastapi
uvicorn
python-multipart
unstructured
langchain-unstructured
faiss-cpu
langchain-google-genai
gunicorn==20.1.0
pydantic==2.9.2
sqlalchemy==2.0.36
mysql-connector-python==9.1.0
langchain-aws
langchain-anthropic
boto3
cryptography==41.0.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
regex==2023.12.25
psycopg2
huggingface_hub
openpyxl
gspread
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib

# requirements for social agent
openai
opencv-python
numpy
Pillow
fastapi
langchain-core

